+++ changelog.md
@@ -11,6 +11,7 @@
   * Minimum number of aircraft for autoplanner to plan OCA packages against 
   * Airbase threat range (nmi)
   * TARCAP threat buffer distance (nmi)
+  * SEAD Sweep threat buffer distance (nmi)
   * AEW&C threat buffer distance (nmi)
   * Theater tanker threat buffer distance (nmi)
 * **[Options]** Improved the option to configure OPFOR autoplanner aggressiveness. The AI might now take even more risks and plan missions against defended targets.